requirements.txt
neat-python==0.92
numpy==1.18.2
pygame==1.9.6
